remote: fix regressions and bugs with REMOTE_PREFETCH_DATA
* Fix a bug with timeShift() or similar functions where the time range
  isn't the one from requestContext (this bug exists in 1.0).
* Do some cleanups.
* Add some TODOs to make the got better later.
