+++ webapp/graphite/render/evaluator.py
@@ -78,27 +78,6 @@ def evaluateTokens(requestContext, tokens, replacements=None):
     raise ValueError("unknown token in target evaluator")
 
 
-def extractPathExpressions(targets):
-  # Returns a list of unique pathExpressions found in the targets list
-
-  pathExpressions = set()
-
-  def extractPathExpression(tokens):
-    if tokens.expression:
-      return extractPathExpression(tokens.expression)
-    elif tokens.pathExpression:
-      pathExpressions.add(tokens.pathExpression)
-    elif tokens.call:
-      for a in tokens.call.args:
-        extractPathExpression(a)
-
-  for target in targets:
-    tokens = grammar.parseString(target)
-    extractPathExpression(tokens)
-
-  return list(pathExpressions)
-
-
 # Avoid import circularities
 from graphite.render.functions import (SeriesFunctions,
                                        NormalizeEmptyResultError)  # noqa
